separate safe wallet code and improve the naming to reduce cogntive load
